% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_vector.R
\name{as_numeric}
\alias{as_numeric}
\title{Converts a vector of elements to numeric elements}
\usage{
as_numeric(input)
}
\arguments{
\item{input}{A vector}
}
\value{
A vector where the numeric elements are numeric and the non-numeric ones are NA
}
\description{
Converts a vector of elements to numeric elements
}
\examples{
as_numeric(c(1,2,3))
as_numeric(c('1',2,3))
as_numeric(c('1','a',3))
as_numeric(c('1','2',NA))
as_numeric(factor(c(1, "4", 3)))
}
