% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_AZRmodel.R
\name{AZRmodel}
\alias{AZRmodel}
\title{AZRmodel class}
\usage{
AZRmodel(input = NULL, simFlag = TRUE)
}
\arguments{
\item{input}{A string with the full filename (including absolute or
relative path) to a text file, describing an AZRmodel.}

\item{simFlag}{If set to TRUE (default) then the model will be prepared for
simulations (dosing, constraints, etc.) and all required sim functions will
be generated and added to the AZRmodel attributes. Choose this option when
you want to simulate. In cases were the model structure and the information
about inputs is important, then choose "FALSE".}
}
\value{
An AZRmodel object. If input is NULL, an empty object is returned.
}
\description{
AZRmodel objects represent dynamic, ODE based models that can be simulated.
The class constructor function AZRmodel() can create such objects from
text file descriptions of these models. If no input argument is provided,
an empty AZRmodel object is returned.
}
\examples{
AZRmodel()
filename <- system.file(package="AZRsim","examples","NovakTyson.txt")
AZRmodel(filename)
}

