% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{AZRxdotcalc}
\alias{AZRxdotcalc}
\title{Return evaluated RHS of AZRmodel ODEs}
\usage{
AZRxdotcalc(model, time = 0, states = NULL, parameters = NULL)
}
\arguments{
\item{model}{An AZRmodel}

\item{time}{Scalar time at which to evaluate the RHS}

\item{states}{Names vector with state values at which to evaluate the ODE RHS
Values for all states in the model need to be provided. If not provided
then the initial conditions in the model will be used. If non-numeric then
potential changes through parameter settings are taken into account.}

\item{parameters}{A named parameter vector to be used for evaluation
Named parameters need to be present in the model.}
}
\value{
Vector with evaluated RHS of ODEs
}
\description{
Calculates and returns the RHS of ODEs in an AZRmodel for given states,
parameters, and time (needs to be scalar). Dosing tables not allowed.
Useful to assess issues in a larger AZRmodel.
}
\examples{
model <- create_model(system.file("examples/NovakTyson.txt", package="AZRsim"))
x <- AZRxdotcalc(model)
x <- AZRxdotcalc(model,400)
x <- AZRxdotcalc(model,400,parameters=c(k1=0.5))
}
