% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{chunk_ode}
\alias{chunk_ode}
\title{chunk an ode into parts split on +/- signs}
\usage{
chunk_ode(.string)
}
\arguments{
\item{.string}{vector of ode strings}
}
\value{
list of chunked odes
}
\description{
chunk an ode into parts split on +/- signs
}
