% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{export_azrmod}
\alias{export_azrmod}
\title{Export of AZRmodel}
\usage{
export_azrmod(model, filename = NULL, useBC = FALSE, useSIM = FALSE)
}
\arguments{
\item{model}{An AZRmodel to be exported.}

\item{filename}{Full path with filename to export the model}

\item{useBC}{Flag for the use of the biochemical notation if TRUE}

\item{useSIM}{Flag to export the model with potential changes for simulation
purposes if TRUE. If FALSE, the originally imported model will be exported.}
}
\value{
None
}
\description{
Export of an AZRmodel to either an ODE based .txt or a biochemical reaction
based .txtbc file. Note that .txtbc files also might contain ODEs, as not all
models allow for the biochemical reaction notation.
}
\examples{
model <- AZRmodel()
export_azrmod(model)
filename <- system.file(package="AZRsim","examples","NovakTyson.txt")
model <- AZRmodel(filename)
export_azrmod(model,"filename")
export_azrmod(model,'filename',useBC=TRUE)
}
