% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{fileparts}
\alias{fileparts}
\title{Get the various parts of a file with path string.}
\usage{
fileparts(filename.with.path)
}
\arguments{
\item{filename.with.path}{A string of a filename with a path}
}
\value{
A list with the following components:
\item{pathname}{The path name}
\item{filename}{The file name}
\item{fileext}{The file extension}
}
\description{
Get the various parts of a file with path string.
}
\examples{
fileparts("a/b/c/d.R")
fileparts("e")
fileparts("f.R")
}
