% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{fileread}
\alias{fileread}
\title{Reads a text file and returns it as string}
\usage{
fileread(filename, collapserows = TRUE)
}
\arguments{
\item{filename}{The filename, possibly including the path}

\item{collapserows}{FALSE: each row of the txt file is returned as a separate list element
TRUE: a single block is returned for the whole text}
}
\value{
Character string with files contents
}
\description{
Should only be used on text files. Will read line by line until the end and
return the results in a character string, using "\\n" as separator between
read lines.
}
\examples{
\dontrun{
filewrite("Hello World!","filename.txt")
fileread("filename.txt")
filewrite("Hello\\nWorld!","filename.txt")
fileread("filename.txt")
fileread("filename.txt",collapserows=FALSE)
}
}

