% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{filewrite}
\alias{filewrite}
\title{Writes a formatted string to a text file}
\usage{
filewrite(text, filename)
}
\arguments{
\item{text}{A character string}

\item{filename, }{possibly including the path}
}
\value{
None
}
\description{
Writes a formatted string to a text file
}
\details{
This is a wrapper to the write function. Ideally, use it only for text to
write to a file. If the file with the filename already exists, it will be
overwritten. If a path is provided and the folder does not exist, the folder
will be created first.
}
\examples{
\dontrun{
filewrite("Hello World!","filename.txt")
filewrite("Hello World!","test/filename.txt")
}
}
