% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{fopen}
\alias{fopen}
\title{Opens a file and creates the folder if needed}
\usage{
fopen(filename, mode = "w")
}
\arguments{
\item{filename}{String with file, possibly including path}

\item{mode}{"w" for writing, "r" for reading}
}
\value{
File ID
}
\description{
Opens a file and creates the folder if needed
}
\examples{
\dontrun{
fopen("e:/test/test.txt")
fopen("../test.R")
}
}

