% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_io.R
\name{fwrite}
\alias{fwrite}
\title{Writes text into an opened file}
\usage{
fwrite(fid, text)
}
\arguments{
\item{fid}{File ID}

\item{text}{Text to write - can be formatted with escaped chars}
}
\value{
None
}
\description{
Writes text into an opened file
}
\details{
File needs to be opened with fopen and closed with fclose
}
\examples{
\dontrun{
fwrite(fid,text)
}
}
