% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{generate_diagram_syntax}
\alias{generate_diagram_syntax}
\title{generate the digraph syntax from a model}
\usage{
generate_diagram_syntax(model)
}
\arguments{
\item{model}{azrmodel}
}
\description{
generate the digraph syntax from a model
}
