% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{get_chunk_signs}
\alias{get_chunk_signs}
\title{determine the signs for each of the chunks in an ode}
\usage{
get_chunk_signs(.string)
}
\arguments{
\item{.string}{vector of strings defining ODEs}
}
\description{
determine the signs for each of the chunks in an ode
}
\examples{
get_chunk_signs("-ka*Ad+F11*input1")
get_chunk_signs(list(
    Ad = "-ka*Ad+F11*input1",
    Ac = "ka*Ad + F22*input2 - CL*Ac/Vc")
)
}
