% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmpt_relationships.R
\name{get_cmpt_relationships}
\alias{get_cmpt_relationships}
\title{get compartment relationships}
\usage{
get_cmpt_relationships(model_states)
}
\arguments{
\item{model_states}{states of of a model}
}
\description{
get compartment relationships
}
