% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{get_inputs}
\alias{get_inputs}
\title{determine the names of the inputs in the ode system of equations}
\usage{
get_inputs(model_states)
}
\arguments{
\item{model_states}{vector of ode model states}
}
\description{
determine the names of the inputs in the ode system of equations
}
\examples{
get_inputs(c("-ka*Ad + F11*input1", "ka*Ad - Cl*Ac/V + F22*input2"))
}
