% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_AZRmodel.R
\name{is.AZRmodel}
\alias{is.AZRmodel}
\title{Check function if it is an AZRmodel}
\usage{
is.AZRmodel(input)
}
\arguments{
\item{input}{AZRmodel object}
}
\value{
TRUE or FALSE
}
\description{
Check function if it is an AZRmodel
}
\examples{
  is.AZRmodel(AZRmodel())
}

