% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_AZRmodel.R
\name{is_azrmod}
\alias{is_azrmod}
\title{Check function if it is an azrmod}
\usage{
is_azrmod(input)
}
\arguments{
\item{input}{azrmod object}
}
\value{
TRUE or FALSE
}
\description{
Check function if it is an azrmod
}
\examples{
  is_azrmod(AZRmodel())
}
