% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_vector.R
\name{isnumericVector}
\alias{isnumericVector}
\title{Checks if a vector consists of numeric entries}
\usage{
isnumericVector(input)
}
\arguments{
\item{input}{A vector}
}
\value{
TRUE if numeric vector, FALSE if not
}
\description{
Checks if a vector has only numeric entries "2" counts as numeric, "a" does
not. "NA" and NA does NOT count as numeric either.
}
\examples{
isnumericVector(c(1,2,3))
isnumericVector(c('1',2,3))
isnumericVector(c('1','a',3))
isnumericVector(c('1','2',NA))
}
