% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{mkdir}
\alias{mkdir}
\title{Creates a folder if it does not yet exist}
\usage{
mkdir(pathdir)
}
\arguments{
\item{pathdir}{String with folder path}
}
\value{
None
}
\description{
Creates a folder if it does not yet exist
}
\examples{
\dontrun{
mkdir("test/test")
mkdir("../test")
}
}
