% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{mkdirp}
\alias{mkdirp}
\title{Creates a folder if it does not yet exist}
\usage{
mkdirp(pathdir)
}
\arguments{
\item{pathdir}{String with folder path}
}
\value{
None
}
\description{
Creates a folder if it does not yet exist
}
\examples{
\dontrun{
mkdirp("test/test")
mkdirp("../test")
}
}
