% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2txt.R
\name{model_to_text}
\alias{model_to_text}
\title{provide the textual representation of a model as a vector of lines}
\usage{
model_to_text(model)
}
\arguments{
\item{model}{azrmodel}
}
\description{
provide the textual representation of a model as a vector of lines
}
\details{
Export of an AZRmodel to a lines of text. ODE representation is used. These
allow a humanly readable format of the model files.
}
