% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.azrsim}
\alias{plot.azrsim}
\title{Plot a simulated model}
\usage{
\method{plot}{azrsim}(object, mfrow_val = NULL, pars = NULL,
  plot_names = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{azrsim}}

\item{mfrow_val}{A vector of length two that describes the dimensions of the lattice plot
(e.g. \code{mfrow = c(2,2)} will create a 2-by-2 grid of plots).}

\item{pars}{A vector of parameter names to plot (as defined in the text file representation of the model).}

\item{plot_names}{A vector of parameter names that will replace the title of each
individual plot in the lattice.}

\item{...}{Additional arguments to \code{plot()}.}
}
\value{
A lattice plot of the variable simulations over time.
}
\description{
Basic plotting functionality for an object of class \code{azrsim}, generated using
the simulation function \code{AZRsim::simulate.azrmod()}.
}
\examples{
sho <- create_model(system.file("examples/sho.txt", package="AZRsim"))
sho_sim <- simulate(sho, 100)
plot(sho_sim, lwd = 2)
}
