% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{position_and_sign}
\alias{position_and_sign}
\title{get the position and assign it a sign value from a str_locate_all output}
\usage{
position_and_sign(.x, sign = "p")
}
\arguments{
\item{.x}{list of start and end positions for locate call}

\item{sign}{the sign value, (p/n)}
}
\description{
get the position and assign it a sign value from a str_locate_all output
}
\examples{
 positives <- stringr::str_locate_all('-ka*Ad+F11*input1', '\\\\+')
 position_and_sign(positives)
}
