% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_AZRmodel.R
\name{print.azrmod}
\alias{print.azrmod}
\title{Generic print function for AZRmodels}
\usage{
\method{print}{azrmod}(x, ...)
}
\arguments{
\item{x}{azrmod object}

\item{...}{Additional unused arguments}
}
\description{
Generic print function for AZRmodels
}
\examples{
  print(azrmod())
}
