% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_AZRmodel.R
\name{print.AZRmodel}
\alias{print.AZRmodel}
\title{Generic print function for AZRmodels}
\usage{
\method{print}{AZRmodel}(x, ...)
}
\arguments{
\item{x}{AZRmodel object}

\item{...}{Additional unused arguments}
}
\description{
Generic print function for AZRmodels
}
\examples{
  print(AZRmodel())
}

