% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{recombine_parens}
\alias{recombine_parens}
\title{recombine equations in parens for one nested layer}
\usage{
recombine_parens(.lv)
}
\arguments{
\item{.lv}{list of vectors of ode chunks}
}
\description{
recombine equations in parens for one nested layer
}
\examples{
recombine_parens(list(c(p = "ka*Ad", p = "(1", n = "F11)*input1")))
}
