% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_helpers.R
\name{sep_values}
\alias{sep_values}
\title{light wrapper for collapse diagrammer separator requirements}
\usage{
sep_values(.values, .sep = ";")
}
\arguments{
\item{.values}{vector to collapse to single string}

\item{.sep}{separator to value}
}
\description{
light wrapper for collapse diagrammer separator requirements
}
