% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_string.R
\name{strexplode}
\alias{strexplode}
\title{Splits a string based on a separator substring}
\usage{
strexplode(input, separator = ",")
}
\arguments{
\item{input}{A string to split}

\item{separator}{A substring to use as separator}
}
\value{
A character vector with the split elements
}
\description{
By default a "," is used as separator.
}
\examples{
strexplode(c("Hello,,Test,,X,,Z"))
strexplode(c("Hello,,Test,,X,,Z"),",")
strexplode(c("Hello,,Test,,X,,Z"),",,")
strexplode(c("Hello,,Test,,X,,Z"),"§")
}

