% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_string.R
\name{strexplodePC}
\alias{strexplodePC}
\title{Splits strings by separator only if separators not in parentheses}
\usage{
strexplodePC(input, separator = ",", group = "round")
}
\arguments{
\item{input}{A string to split}

\item{separator}{A substring to use as separator}

\item{group}{"round", "square", or "curly" allowing grouping by parantheses
of defined type in which elements are not exploded}
}
\value{
A character vector with the split elements
}
\description{
By default a "," is used as separator.
}
\examples{
strexplodePC("(a,b),(c,d)")
strexplodePC("((a,b),(c,d))") # compare with previous - lowest level considered
strexplodePC("(a,b),(c,d))")
strexplodePC("(a;b);(c;d)",";")
}
