% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_string.R
\name{strlocateall}
\alias{strlocateall}
\title{Returns indices of occurrence of strings in strings}
\usage{
strlocateall(input, searchString)
}
\arguments{
\item{input}{A string to search in}

\item{searchString}{A string to search for}
}
\value{
A list with the following components:
\item{start}{A vector with start indices}
\item{end}{A vector with end indices of matching}
}
\description{
Returns indices of occurrence of strings in strings
}
\examples{
strlocateall("abcdefgabcdefg","a")
strlocateall("abcdefgabcdefg","ab")
strlocateall("abcdefgabcdefg","x")
}

