% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_string.R
\name{strmatch}
\alias{strmatch}
\title{Returns indices of elements in a vector of that match a search string}
\usage{
strmatch(inputVec, searchString)
}
\arguments{
\item{inputVec}{A vector of strings}

\item{searchString}{A string to search for (case sensitive,
exact matching) in inputVec}
}
\value{
A vector with indices of elements that have an exact match.
  NULL is returned if string not found
}
\description{
Exact, case sensitive matching is used.
Function can also be used for other objects than strings.
}
\examples{
strmatch(c("hello","test","a","e","hello"),"hello")
strmatch(c("hello","test","a","e","hello"),"Hello")
strmatch(c("hello","test","a","e","hello"),"xyz")
strmatch(c(1,2,3,4,8,3),3)
}
