% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_string.R
\name{strremWhite}
\alias{strremWhite}
\title{Removes all white spaces in a string}
\usage{
strremWhite(inputText)
}
\arguments{
\item{inputText}{A character string}
}
\value{
Character string with whitespaces removed
}
\description{
Whitespace characters are the following: intToUtf8(c(9, 10, 11, 12, 13, 32))
}
\examples{
strremWhite("abcde")
strremWhite("a b c d e")
strremWhite(" a b c d e    ")
strremWhite(intToUtf8(c(9, 10, 11, 12, 13, 32, 97,  32,  98,  32,  99,  32, 100,  32, 101, 32, 13)))
}
