% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_string.R
\name{strrepM}
\alias{strrepM}
\title{String replacement function}
\usage{
strrepM(origstr, oldsubstr, newsubstr)
}
\arguments{
\item{origstr}{A character string in which to replace sub strings}

\item{oldsubstr}{Search string to replace}

\item{newsubstr}{A string to replace the search string if found}
}
\value{
Resulting output string
}
\description{
String replacement function
}
\examples{
strrepM("Hello World!","World","Kitty")
strrepM("a b c d e"," ","")
}
