% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_string.R
\name{strtrimM}
\alias{strtrimM}
\title{Truncates strings by removing leading and trailing white spaces}
\usage{
strtrimM(input)
}
\arguments{
\item{input}{A character string}
}
\value{
Character string with leading and trailing whitespaces removed
}
\description{
Whitespace characters are the following: intToUtf8(c(9, 10, 11, 12, 13, 32))
}
\examples{
strtrimM("abcde")
strtrimM("a b c d e")
strtrimM(" a b c d e    ")
strtrimM(intToUtf8(c(9, 10, 11, 12, 13, 32, 97,  32,  98,  32,  99,  32, 100,  32, 101, 32, 13)))
}
