% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2model.R
\name{text_to_model}
\alias{text_to_model}
\title{convert vector of strings to a model object}
\usage{
text_to_model(modelText)
}
\arguments{
\item{modelText}{strings of model file}
}
\description{
convert vector of strings to a model object
}
