% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2txt.R
\name{write_model}
\alias{write_model}
\title{export model to a file}
\usage{
write_model(model, .file)
}
\arguments{
\item{model}{azrmodel object}

\item{.file}{path to write}
}
\description{
export model to a file
}
