% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{azrmod class}
\usage{
create_model(input = NULL, simFlag = TRUE)
}
\arguments{
\item{input}{A string with the full filename (including absolute or
relative path) to a text file, describing an azrmod.}

\item{simFlag}{If set to \code{TRUE} (default) then the model will be prepared for
simulations (dosing, constraints, etc.) and all required sim functions will
be generated and added to the azrmod attributes. Choose this option when
you want to simulate. In cases were the model structure and the information
about inputs is important, then choose \code{FALSE}.}
}
\value{
An object of class azr_model. If input is NULL, an empty object is returned.
}
\description{
This function creates an object of class \code{azrmod} which represents dynamic,
ODE based models that can be simulated. The class constructor function
\code{azrmod} creates such objects from specifc text file descriptions of
these models. If no input argument is provided, an empty azrmod object is returned.
}
\examples{
empty_mod <- create_model()
filename <- system.file(package="AZRsim","examples","NovakTyson.txt")
mod <- create_model(filename)
}
