% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries_io.R
\name{rmdir}
\alias{rmdir}
\title{Removes a folder}
\usage{
rmdir(pathdir)
}
\arguments{
\item{pathdir}{String with folder path}
}
\value{
None
}
\description{
Removes a folder
}
\examples{
\dontrun{
rmdir("test/test")
rmdir("../test")
}
}

