% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gui.R
\name{shiny_plot}
\alias{shiny_plot}
\title{Use shiny to graphically represent the contents of a dataframe}
\usage{
shiny_plot(data)
}
\arguments{
\item{data}{An object of class azr_simulate.}
}
\value{
Launches an interactive Shiny interface of all the parameters in the
default web browser.
}
\description{
As input data frames are used and the columns plotted.
By default it looks for a "time" to be used as X-axis.
In the future it might handle "ID" columns in some way and others,
such as "TAD" and "AMT", ... Lets start simple for now.
Only numeric columns in the dataframe are considered.
}
\examples{
\dontrun{
model <- create_model(system.file("examples/NovakTyson.txt", package="AZRsim"))
x <- simulate(model,400)
shiny_plot(x)
}
}
